requirements.txt
requests==2.24.0
pefile==2019.4.18
capstone==4.0.2
python_magic==0.4.18
termcolor==1.1.0
